Remove `join_context_async`, since there is no way to use the `injected` boolean (explanation in commit message)
`injected` boolean isn't used anymore, since it's not possible to update
the injected state of the future once it has been created. In
`join_context`, `injected` represents if a different thread is executing
the function than the thread it was created on. In `join_async` a future
can indeed be executed by a different thread, but unfortunately we have
no way of telling an already started future that it is now being
executed by a different thread, so we just don't support the `injected`
boolean.