Fix bug where we tried to perform a non-atomic random deque selection and subsequent removal
When we choose a random index into the stealable set, it's possible that
another thread has chosen this same index as well. This means if one
thread successfully steals the entire deque and removes the ID
corresponding to the same index from the respective stealable set, by
the time the other thread goes to retrieve the deque ID from the
stealable set it will no longer be there. This is why we cannot unwrap,
and must instead retry (somehow, still need to see if accesses and
mutations to stealable set need to truly be atomic or not, if there's a
retry mechanism if the same victim thread can be kept or if a new random
thread needs to be chosen, etc.).