Fix bug where new deques were not being added to underlying stealable sets
New deques were only being added to the Stealables mapping, not the
underlying stealable sets. This meant that some deques could never be
found and stolen by other threads, since these deques could not be
returned by querying a stealable set for a random deque (since these
deques were not in a stealable set). This means it was possible to
encounter a deadlock, like in the `join_context_both` test case, where a
worker thread that is executing a blocking job can only be unblocked by
having another job on its deque be stolen and executed by another
thread. Since the deque was never added to the stealable set, it could
never be found and popped off from by another worker thread, and the
worker thread would be stuck forever.